apps/tappweb/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "tappweb" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../coverage/apps/tappweb/html' --cov-report xml:'../../coverage/apps/tappweb/coverage.xml' --html='../../reports/apps/tappweb/unittests/html/index.html' --junitxml='../../reports/apps/tappweb/unittests/junit.xml'"

[tool.poetry]
name = "tappweb"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [tool.poetry.dependencies]
  python = "^3.11"
  fastapi = "^0.97.0"
  sqlalchemy = { extras = ["asyncio"], version = "^2.0.16" }
  uvicorn = { extras = ["standard"], version = "^0.22.0" }
  alembic = "^1.11.1"
  asyncpg = "^0.27.0"
  aiohttp = "^3.8.4"
  apscheduler = "^3.10.1"
  pyjwt = "^2.6.0"
  fastapi-pagination = "^0.12.4"
  gunicorn = "^20.1.0"
  typer = { extras = ["all"], version = "^0.9.0" }
  greenlet = "^2.0.2"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.0.2"
  flake8 = "6.0.0"
  pytest = "7.3.1"
  pytest-sugar = "0.9.7"
  pytest-cov = "4.1.0"
  pytest-html = "3.2.0"
  pre-commit = "^3.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.nx]
autoActivate = true

[tool.black]
line-length = 120
skip-magic-trailing-comma = true
verbose = true

[tool.ruff]
line-length = 120

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
fail-under = 100
verbose = 2